refactor: remove unused pr action stub
